backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pillow>=10.0.0
google-generativeai>=0.3.0
passlib[bcrypt]>=1.7.0
python-jose[cryptography]>=3.3.0
requests>=2.31.0 